requirements.txt
Flask==1.1.1
jinja2==3.0.3
Flask_SQLAlchemy==2.5.1
psycopg2_binary==2.9.3